fixed: spec fail on ruby1.8.7
